Refactor calibration view
