+++ Sources/ArcGISToolkit/Components/Augmented Reality/WorldScaleGeoTrackingSceneView.swift
@@ -20,6 +20,8 @@ import ArcGIS
 public struct WorldScaleGeoTrackingSceneView: View {
     /// The proxy for the ARSwiftUIView.
     @State private var arViewProxy = ARSwiftUIViewProxy()
+    /// The scene view.
+    @State private var scene: ArcGIS.Scene
     /// The camera controller that will be set on the scene view.
     @State private var cameraController: TransformationMatrixCameraController
     /// The current interface orientation.
@@ -48,32 +50,15 @@ public struct WorldScaleGeoTrackingSceneView: View {
     @State private var currentLocation: Location?
     /// The valid accuracy threshold for a location in meters.
     private var validAccuracyThreshold = 0.0
-    /// The basemap opacity.
-    @State private var basemapOpacity = 0.0
     /// A Boolean value that indicates if the AR experince is being calibrated.
     @State private var isCalibrating = false
-    /// The slider value for the camera heading.
-    @State private var headingSliderValue = 0.0
-    /// The slider value for the camera elevation.
-    @State private var elevationSliderValue = 0.0
-    /// The elevation timer for the "joystick" behavior.
-    @State private var headingTimer: Timer?
-    /// The heading timer for the "joystick" behavior.
-    @State private var elevationTimer: Timer?
-    /// The heading delta amount based on the heading slider value.
-    private var joystickHeadingDelta: Double {
-        Double(signOf: headingSliderValue, magnitudeOf: headingSliderValue * headingSliderValue / 25)
-    }
-    /// The elevation delta amount based on the elevation slider value.
-    private var joystickElevationDelta: Double {
-        Double(signOf: elevationSliderValue, magnitudeOf: elevationSliderValue * elevationSliderValue / 100)
-    }
     
     /// Creates a world scale scene view.
     /// - Parameters:
     ///   - locationDataSource: The location datasource used to acquire the device's location.
     ///   - clippingDistance: Determines the clipping distance in meters around the camera. A value
     ///   of `nil` means that no data will be clipped.
+    ///   - scene: The scene used in the scene view.
     ///   - sceneView: A closure that builds the scene view to be overlayed on top of the
     ///   augmented reality video feed.
     /// - Remark: The provided scene view will have certain properties overridden in order to
@@ -82,8 +67,12 @@ public struct WorldScaleGeoTrackingSceneView: View {
     public init(
         locationDataSource: LocationDataSource = SystemLocationDataSource(),
         clippingDistance: Double? = nil,
+        scene: ArcGIS.Scene,
         @ViewBuilder sceneView: @escaping (SceneViewProxy) -> SceneView
     ) {
+        scene.basemap?.baseLayers.forEach({ $0.opacity = 0.5})
+        _scene = .init(initialValue: scene)
+        
         sceneViewBuilder = sceneView
         
         let cameraController = TransformationMatrixCameraController()
@@ -170,10 +159,20 @@ public struct WorldScaleGeoTrackingSceneView: View {
                 }
             } catch {}
         }
+        .onChange(of: scene.basemap?.loadStatus) { loadStatus in
+            guard loadStatus == .loaded else { return }
+            setBasemapOpacity(0)
+        }
         .toolbar {
             ToolbarItem(placement: .bottomBar) {
                 if !calibrationViewIsHidden {
-                    calibrationView
+                    CalibrationView(
+                        scene: scene,
+                        cameraController: cameraController,
+                        isCalibrating: isCalibrating,
+                        calibrationHeading: calibrationHeading,
+                        calibrationElevation: calibrationElevation
+                    )
                 }
             }
         }
@@ -282,123 +281,11 @@ public struct WorldScaleGeoTrackingSceneView: View {
         return view
     }
     
-    /// Updates the heading of the scene view camera controller.
-    /// - Parameter heading: The camera heading.
-    func updateHeading(_ heading: Double) {
-        cameraController.originCamera = cameraController.originCamera.rotatedTo(
-            heading: heading,
-            pitch: cameraController.originCamera.pitch,
-            roll: cameraController.originCamera.roll
-        )
-        calibrationHeading = heading
-    }
-    
-    /// Rotates the heading of the scene view camera controller by the heading delta in degrees.
-    /// - Parameter headingDelta: The heading delta in degrees.
-    func rotateHeading(_ headingDelta: Double) {
-        let newHeading = cameraController.originCamera.heading + headingDelta
-        cameraController.originCamera = cameraController.originCamera.rotatedTo(
-            heading: newHeading,
-            pitch: cameraController.originCamera.pitch,
-            roll: cameraController.originCamera.roll
-        )
-        calibrationHeading = newHeading
-    }
-    
-    /// Elevates the scene view camera controller by the elevation delta.
-    /// - Parameter elevationDelta: The elevation delta.
-    func updateElevation(_ elevationDelta: Double) {
-        cameraController.originCamera = cameraController.originCamera.elevated(by: elevationDelta)
-        if let elevation = cameraController.originCamera.location.z {
-            calibrationElevation = elevation
-        }
-    }
-    
-    /// A view that allows the user to calibrate the heading of the scene view camera controller.
-    var calibrationView: some View {
-        Button {
-            isCalibrating = true
-            basemapOpacity = 0.5
-        } label: {
-            Text("Calibrate")
-        }
-        .popover(isPresented: $isCalibrating) {
-            VStack {
-                Text("Heading: \(calibrationHeading?.rounded(.towardZero) ?? cameraController.originCamera.heading.rounded(.towardZero), format: .number)")
-                
-                HStack {
-                    Button {
-                        let heading = cameraController.originCamera.heading - 1
-                        updateHeading(heading)
-                    } label: {
-                        Image(systemName: "minus")
-                    }
-                    Slider(value: $headingSliderValue, in: -10...10) { editingChanged in
-                        if !editingChanged {
-                            headingTimer?.invalidate()
-                            headingTimer = nil
-                            headingSliderValue = 0.0
-                        }
-                    }
-                    .onChange(of: headingSliderValue) { heading in
-                        
-                        guard headingTimer == nil else { return }
-                        // Create a timer which rotates the camera when fired.
-                        let timer = Timer(timeInterval: 0.1, repeats: true) { [self] (_) in
-                            rotateHeading(joystickHeadingDelta)
-                        }
-                        headingTimer = timer
-                        // Add the timer to the main run loop.
-                        RunLoop.main.add(timer, forMode: .default)
-                    }
-                    
-                    Button {
-                        let heading = cameraController.originCamera.heading + 1
-                        updateHeading(heading)
-                    } label: {
-                        Image(systemName: "plus")
-                    }
-                }
-                VStack {
-                    Text("Elevation: \(calibrationElevation?.rounded(.towardZero) ?? cameraController.originCamera.location.z?.rounded(.towardZero) ?? 0, format: .number)")
-                    HStack {
-                        Button {
-                            updateElevation(-1)
-                        } label: {
-                            Image(systemName: "minus")
-                        }
-                        Slider(value: $elevationSliderValue, in: -20...20) { editingChanged in
-                            if !editingChanged {
-                                elevationTimer?.invalidate()
-                                elevationTimer = nil
-                                elevationSliderValue = 0.0
-                            }
-                        }
-                        .onChange(of: elevationSliderValue) { elevation in
-                            
-                            guard elevationTimer == nil else { return }
-                            // Create a timer which rotates the camera when fired.
-                            let timer = Timer(timeInterval: 0.1, repeats: true) { [self] (_) in
-                                updateElevation(joystickElevationDelta)
-                            }
-                            elevationTimer = timer
-                            // Add the timer to the main run loop.
-                            RunLoop.main.add(timer, forMode: .default)
-                        }
-                        Button {
-                            updateElevation(1)
-                        } label: {
-                            Image(systemName: "plus")
-                        }
-                    }
-                }
-            }
-            .frame(minWidth: 200, maxHeight: /*@START_MENU_TOKEN@*/.infinity/*@END_MENU_TOKEN@*/)
-            .padding()
-        }
-        .onDisappear {
-            basemapOpacity = 0.0
-        }
+    /// Sets the basemap base layers with the given opacity.
+    /// - Parameter opacity: The opacity of the layer.
+    private func setBasemapOpacity(_ opacity: Float) {
+        guard let basemap = scene.basemap else { return }
+        basemap.baseLayers.forEach { $0.opacity = opacity }
     }
     
     /// A view that displays the horizontal and vertical accuracy of the current location datasource location.
